Started the remix example, but noticed a lot of issues
I added a lot of TODOs to fix those issues in the next iteration. Before we finish the remix example, we should redo how we load and store data for the frontend. That way, it'll be less magical and controlled by users. We should not assume they use a URL router outside their server.